+++ CHANGELOG.md
@@ -2,6 +2,42 @@
 
 ## Version 20
 
+### v20.8.0
+
+- Feat: providing child logger to `beforeRouting()` hook:
+  - The function assigned to config property `server.beforeRouting` now accepts additional argument `getChildLogger()`;
+  - The featured method accepts `request` and returns a child logger if `childLoggerProvider()` is configured;
+  - Otherwise, it returns the root logger (same for all requests, same as the `logger` argument).
+
+```ts
+import { createConfig } from "express-zod-api";
+import { randomUUID } from "node:crypto";
+
+const config = createConfig({
+  logger: { level: "debug" },
+  childLoggerProvider: ({ parent }) =>
+    parent.child({ requestId: randomUUID() }),
+  server: {
+    listen: 80,
+    beforeRouting: ({ app, logger, getChildLogger }) => {
+      logger.info("This is root logger");
+      app.use((req, res, next) => {
+        getChildLogger(req).info("This is a child logger");
+        next();
+      });
+    },
+  },
+});
+```
+
+### v20.7.1
+
+- Improved documentation on error handling:
+  - More clarity on the origins of possible runtime errors and how they are handled by default;
+  - Revealing details on how routing, parsing and upload errors are handled by default;
+  - Correction to the JSDoc of the corresponding `errorHandler` property in config.
+- Removing redundant type coercion in the migration tool.
+
 ### v20.7.0
 
 - Changes to migration plugin (single-use tool, regardless SemVer):